파일 구조 변경 & 투표 로직 구조 디자인 완성 & 예약 시스템 도입

+++ src/main/java/service/chat/mealmate/mealmate/repository/VoteRepository.java
@@ -2,8 +2,8 @@ package service.chat.mealmate.mealmate.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import service.chat.mealmate.mealmate.domain.Vote;
+import service.chat.mealmate.mealmate.domain.vote.Vote;
 
 @Repository
-public interface VoteRepository extends JpaRepository<Vote, Integer>{
+public interface VoteRepository extends JpaRepository<Vote, Long>{
 }
